requirements.txt
# FraudLens AI — NVIDIA GTC 2026
# Streamlit + NVIDIA NIM multi-agent fraud detection

# Web UI
streamlit>=1.30.0
plotly>=5.18.0

# NVIDIA NIM Client (OpenAI-compatible)
openai>=1.0.0
httpx>=0.25.0

# PDF & Image Processing
PyMuPDF>=1.23.0
pypdf>=3.0.0
Pillow>=10.0.0

# Async
aiohttp>=3.9.0

# Environment
python-dotenv>=1.0.0

# Logging
loguru>=0.7.0
